+++ apiserver/plane/app/views/__init__.py
@@ -30,7 +30,7 @@ from .user.base import (
 
 from .oauth import OauthEndpoint
 
-from .base import BaseAPIView, BaseViewSet, WebhookMixin
+from .base import BaseAPIView, BaseViewSet
 
 from .workspace.base import (
     WorkSpaceViewSet,
